Rename and move into a package the FileUpload
