+++ app/src/main/java/com/ncnf/event/create/EventCreateActivity.java
@@ -33,7 +33,7 @@ import com.ncnf.event.PublicEvent;
 import com.ncnf.main.MainActivity;
 import com.ncnf.user.User;
 import com.ncnf.utilities.DateAdapter;
-import com.ncnf.utilities.FileUpload;
+import com.ncnf.storage.FileStore;
 import com.ncnf.utilities.InputValidator;
 
 import java.text.DecimalFormat;
@@ -188,7 +188,7 @@ public class EventCreateActivity extends AppCompatActivity implements AdapterVie
                         // TODO: might change with the new event class
                         // File upload must happened after the event is saved, such that the image path
                         // contains the event's UUID
-                        FileUpload file = new FileUpload(Event.IMAGE_PATH, String.format(Event.IMAGE_NAME, "PLEASE_REPLACE_WITH_UUID"));
+                        FileStore file = new FileStore(Event.IMAGE_PATH, String.format(Event.IMAGE_NAME, "PLEASE_REPLACE_WITH_UUID"));
                         pictureView.setDrawingCacheEnabled(true);
                         pictureView.buildDrawingCache();
                         Bitmap bitmap = ((BitmapDrawable) pictureView.getDrawable()).getBitmap();
